Replacing size_t data type with int
removed using size_t data type and replacing them with int data type for variable equal_pos in DELETE and UPDATE function

+++ assignment.cpp
@@ -90,8 +90,8 @@ void selectcommand(vector<string> selectcommand){ //SELECT
 
 void Updatecommand(vector<string> Updatecommand){
 
-    size_t equal_pos1;             //uses size_t data, which is an unsigned integer type
-    size_t equal_pos2;            
+    int equal_pos1;             //uses size_t data, which is an unsigned integer type
+    int equal_pos2;            
     int customer_id;
     int WHERE_index;
     int update_index;
@@ -154,7 +154,7 @@ void Updatecommand(vector<string> Updatecommand){
 
 void Deletecommand(vector<string> Deletecommand){
 
-    size_t equal_pos; //uses size_t data
+    int equal_pos; //uses size_t data
     string extracted_id;
     string del_location;
     string comparer;
